chore: move selection about code
